+++ R/CartoPred.R
@@ -454,3 +454,29 @@ Map_predict <- function(object, saveWD, mask = NULL, Num = NULL,
   }
   return(pred_map)
 } # end of function
+
+#' Transform raster as data.frame to be later used with ggplot
+#' Modified from rasterVis::gplot
+#'
+#' @param x A Raster* object
+#' @param maxpixels Maximum number of pixels to use
+#'
+#' @details rasterVis::gplot is nice to plot a raster in a ggplot but
+#' if you want to plot different rasters on the same plot, you are stuck.
+#' If you want to add other information or transform your raster as a
+#' category raster, you can not do it. With `SDMSelect::gplot_data`, you retrieve your
+#' raster as a data.frame that can be modified as wanted using `dplyr` and
+#' then plot in `ggplot` using `geom_tile`.
+#'
+#' @export
+
+gplot_data <- function(x, maxpixels = 50000)  {
+  x <- raster::sampleRegular(x, maxpixels, asRaster = TRUE)
+  coords <- raster::xyFromCell(x, seq_len(raster::ncell(x)))
+  ## Extract values
+  dat <- raster::stack(as.data.frame(raster::getValues(x)))
+  names(dat) <- c('value', 'variable')
+
+  dat <- dplyr::as.tbl(cbind(coords, dat))
+  dat
+}